requirements.txt
os
argparse
shutil
gzip
six.moves
numpy
matplotlib
tensorflow-gpu>=1.6
